Add examples for async as test cases
This includes execution flow similar to:
* Actor ask/tell
* Promises with callbacks
* Work interleaved on a thread using suspend/resume.

The implementations of these execution models are obviously very simplistic, but intended to emphasize the tracing aspect.

These tests also make heavy use of the Phaser in order to insure deterministic execution of the tests for multi-threaded execution.  Hopefully this does not detract from the examples too much.